Use updated alphaess pypy library version 0.0.3 (supports more detailed exceptions)
Addressed pull comments
https://github.com/home-assistant/core/pull/62801#discussion_r776386569
https://github.com/home-assistant/core/pull/62801#discussion_r776398169
https://github.com/home-assistant/core/pull/62801#discussion_r776398806
https://github.com/home-assistant/core/pull/62801#discussion_r776399796
