+++ homeassistant/components/alphaess/coordinator.py
@@ -1,20 +1,16 @@
 """Coordinator for AlphaEss integration."""
 
-from datetime import timedelta
 import logging
 
+import aiohttp
 from alphaess import alphaess
 
-from homeassistant.config_entries import ConfigEntry
-from homeassistant.const import Platform
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, UpdateFailed
 
-from .const import DOMAIN
+from .const import DOMAIN, SCAN_INTERVAL
 
 _LOGGER: logging.Logger = logging.getLogger(__package__)
-SCAN_INTERVAL = timedelta(minutes=5)
-PLATFORMS = [Platform.SENSOR]
 
 
 class AlphaESSDataUpdateCoordinator(DataUpdateCoordinator):
@@ -31,18 +27,8 @@ class AlphaESSDataUpdateCoordinator(DataUpdateCoordinator):
         try:
             _LOGGER.info("Trying to query AlphaESS api data")
             return await self.api.getdata()
-        except Exception as exception:
-            raise UpdateFailed() from exception
-
-
-async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
-    """Unload a config entry."""
-    if unload_ok := await hass.config_entries.async_unload_platforms(entry, PLATFORMS):
-        hass.data[DOMAIN].pop(entry.entry_id)
-
-    return unload_ok
-
-
-async def async_update_options(hass: HomeAssistant, entry: ConfigEntry) -> None:
-    """Update options."""
-    await hass.config_entries.async_reload(entry.entry_id)
+        except (
+            aiohttp.client_exceptions.ClientConnectorError,
+            aiohttp.ClientResponseError,
+        ) as error:
+            raise UpdateFailed(error) from error
