+++ homeassistant/components/alphaess/manifest.json
@@ -4,7 +4,7 @@
   "config_flow": true,
   "documentation": "https://www.home-assistant.io/integrations/alphaess",
   "requirements": [
-    "alphaess==0.0.2"
+    "alphaess==0.0.3"
   ],
   "codeowners": [
     "@CharlesGillanders"
