+++ requirements_all.txt
@@ -290,7 +290,7 @@ aladdin_connect==0.3
 alpha_vantage==2.3.1
 
 # homeassistant.components.alphaess
-alphaess==0.0.2
+alphaess==0.0.3
 
 # homeassistant.components.ambee
 ambee==0.4.0
